35강. 애플리케이션 대신 DB로 기능 구현하기

+++ src/test/kotlin/edgar/study/kotlin/libraryapp/service/book/BookServiceTest.kt
@@ -167,7 +167,7 @@ class BookServiceTest @Autowired constructor(
     private fun getBookStatTestThenCount(
         bookStats: List<BookStatResponse>,
         type: BookType,
-        count: Int) {
+        count: Long ) {
         assertThat(bookStats.first { res -> res.type == type }.count).isEqualTo(count)
     }
 }
